image update added

+++ src/main/java/com/soumyajit/E_Grocery/Shop/DTOS/CartItemDTO.java
@@ -2,12 +2,14 @@ package com.soumyajit.E_Grocery.Shop.DTOS;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CartItemDTO {
     private Long id;
     private Long productId;
     private int quantity;
-    private double totalQuantity;
-    private double totalPrice;
+    private BigDecimal totalQuantity;
+    private BigDecimal totalPrice;
     private String unitLabel;
 }
